+++ sample/app/ui-material/material-sample-form.model.ts
@@ -11,15 +11,7 @@ import {
 import { customDateRangeValidator } from "../app.validators";
 
 export const MATERIAL_SAMPLE_FORM_MODEL = [
-    /*
-    new DynamicInputModel({
 
-        id: "chips",
-        placeholder: "Chips",
-        multiple: true,
-        value: ["Hello", "Chips"]
-    }),
-    */
     new DynamicFormGroupModel({
 
         id: "stay",
@@ -56,6 +48,7 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
 
                 id: "roomSize",
                 placeholder: "Room Size",
+                hint: "Choose a room type",
                 options: [
                     {
                         label: "Single Room",
@@ -245,6 +238,14 @@ export const MATERIAL_SAMPLE_FORM_MODEL = [
         id: "note",
         rows: 3,
         placeholder: "Personal Note"
+    }),
+
+    new DynamicInputModel({
+
+        id: "tags",
+        placeholder: "Tags",
+        multiple: true,
+        value: ["hotel", "booking"]
     }),
 
     new DynamicSwitchModel({